refactor(install): extract worker_llm wizard to installer/wizards/ for shared use
Move the ~300-line worker-LLM prompt flow out of
spellbook/cli/commands/install.py into installer/wizards/worker_llm.py as
run_worker_llm_wizard(args). The old entry point retains a thin compat shim
so existing tests and external callers continue to work.

This unblocks wiring the wizard from the root install.py (curl-pipe) entry
path without duplicating ~300 lines of prompt logic.
